execution/requirements.txt
python-binance==1.0.19
websockets==12.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
sqlalchemy==2.0.30
python-dotenv==1.0.1
aiohttp==3.9.5
